backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
httpx>=0.24.0
python-multipart>=0.0.6
